SHOP/requirements.txt
asgiref==3.2.3
boto==2.49.0
boto3==1.11.12
botocore==1.14.12
Django==3.0.2
django-debug-toolbar==2.2
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
